+++ server/controller/src/main/java/ai/starwhale/mlops/api/protocol/runtime/DeviceVO.java
@@ -5,7 +5,7 @@
  * in accordance with the terms of the license agreement you entered into with StarWhale.com.
  */
 
-package ai.starwhale.mlops.api.protocol.agent;
+package ai.starwhale.mlops.api.protocol.runtime;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
